requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pymupdf==1.23.7
langchain==0.0.335
qdrant-client==1.6.4
openai==1.3.5
pytest==7.4.3
httpx==0.25.1
pydantic==2.4.2
python-dotenv==1.0.0
pydantic-settings
